clean up dsf crate; move IdAlloc to ids crate
