+++ dsf/src/tracked_union_find.rs
@@ -1,28 +1,29 @@
-#![allow(missing_docs)]
-#![allow(clippy::type_complexity)]
-use std::{cmp::Reverse, collections::VecDeque, mem::ManuallyDrop, sync::Arc};
+//! A `TrackedUnionFind` augments a [`UnionFind`] structure with change tracking.
+use std::{cmp::Reverse, collections::VecDeque, iter::FusedIterator, mem::ManuallyDrop, sync::Arc};
 
-use atomic::Atomic;
-use bytemuck::NoUninit;
-use imctk_ids::{id_vec::IdVec, Id, Id64};
+use imctk_ids::{id_vec::IdVec, Id, Id64, IdAlloc};
 use priority_queue::PriorityQueue;
 
-use crate::union_find::{Element, UnionFind};
+use crate::{Element, UnionFind};
 
+#[cfg(test)]
+#[path = "tests/test_tracked_union_find.rs"]
+mod test_tracked_union_find;
+
+/// Globally unique ID for a `TrackedUnionFind`.
 #[derive(Id, Debug)]
 #[repr(transparent)]
 pub struct TrackedUnionFindId(Id64);
-// SAFETY: trust me bro
-unsafe impl NoUninit for TrackedUnionFindId {}
+/// Generation number for a `TrackedUnionFind`
 #[derive(Id, Debug)]
 #[repr(transparent)]
 pub struct Generation(u64);
+/// Observer ID for a `TrackedUnionFind` (not globally unique)
 #[derive(Id, Debug)]
 #[repr(transparent)]
 pub struct ObserverId(Id64);
-// SAFETY: trust me bro
-unsafe impl NoUninit for ObserverId {}
 
+/// An `ObserverToken` represents an observer of a `TrackedUnionFind`.
 #[derive(Debug)]
 pub struct ObserverToken {
     tuf_id: TrackedUnionFindId,
@@ -30,6 +31,44 @@ pub struct ObserverToken {
     observer_id: ObserverId,
 }
 
+impl ObserverToken {
+    /// Returns the ID of the associated `TrackedUnionFind`.
+    pub fn tuf_id(&self) -> TrackedUnionFindId {
+        self.tuf_id
+    }
+    /// Returns the ID of the generation that this observer is on.
+    pub fn generation(&self) -> Generation {
+        self.generation
+    }
+    /// Returns the ID of the observer.
+    ///
+    /// Note that observer IDs are local to each `TrackedUnionFind`.
+    pub fn observer_id(&self) -> ObserverId {
+        self.observer_id
+    }
+    /// Returns `true` iff `self` and `other` belong to the same `TrackedUnionFind`.
+    ///
+    /// NB: This does **not** imply that variable IDs are compatible, for that you want `is_compatible`.
+    pub fn is_same_tuf(&self, other: &ObserverToken) -> bool {
+        self.tuf_id == other.tuf_id
+    }
+    /// Returns `true` iff `self` and `other` have compatible variable IDs.
+    ///
+    /// This is equivalent to whether they are on the same generation of the same `TrackedUnionFind`.
+    pub fn is_compatible(&self, other: &ObserverToken) -> bool {
+        self.tuf_id == other.tuf_id && self.generation == other.generation
+    }
+}
+
+/// `Renumbering` represents a renumbering of all variables in `TrackedUnionFind`.
+///
+/// A renumbering stores a forward and a reverse mapping, and the old and new generation IDs.
+///
+/// The forward mapping maps each old variable to an optional new variable (variables may be deleted).
+/// It is a requirement that equivalent old variables are either both mapped to the same new variable or both deleted.
+///
+/// The reverse mapping maps each new variable to its old representative.
+/// The new set of variables is required to be contiguous, hence `reverse` is a total mapping.
 pub struct Renumbering<Atom, Elem> {
     forward: IdVec<Atom, Option<Elem>>,
     reverse: IdVec<Atom, Elem>,
@@ -48,8 +87,12 @@ impl<Atom: Id, Elem: Id> std::fmt::Debug for Renumbering<Atom, Elem> {
     }
 }
 
-impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> Renumbering<Atom, Elem> {
-    pub fn get_reverse(forward: &IdVec<Atom, Option<Elem>>, union_find: &UnionFind<Atom, Elem>) -> IdVec<Atom, Elem> {
+impl<Atom: Id, Elem: Id + Element<Atom>> Renumbering<Atom, Elem> {
+    /// Returns the inverse of a reassignment of variables.
+    pub fn get_reverse(
+        forward: &IdVec<Atom, Option<Elem>>,
+        union_find: &UnionFind<Atom, Elem>,
+    ) -> IdVec<Atom, Elem> {
         let mut reverse: IdVec<Atom, Option<Elem>> = IdVec::default();
         for (old, &new_opt) in forward {
             if let Some(new) = new_opt {
@@ -60,6 +103,7 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> Renumbering<Atom, Elem> {
         }
         IdVec::from_vec(reverse.iter().map(|x| x.1.unwrap()).collect())
     }
+    /// Returns `true` iff the arguments are inverses of each other.
     pub fn is_inverse(forward: &IdVec<Atom, Option<Elem>>, reverse: &IdVec<Atom, Elem>) -> bool {
         reverse.iter().all(|(new, &old)| {
             if let Some(&Some(e)) = forward.get(old.atom()) {
@@ -69,6 +113,7 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> Renumbering<Atom, Elem> {
             }
         })
     }
+    /// Creates a renumbering without checking whether the arguments are valid.
     pub fn new_unchecked(
         forward: IdVec<Atom, Option<Elem>>,
         reverse: IdVec<Atom, Elem>,
@@ -82,6 +127,7 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> Renumbering<Atom, Elem> {
             new_generation,
         }
     }
+    /// Creates a new renumbering from the given forward and reverse assignment.
     pub fn new(
         forward: IdVec<Atom, Option<Elem>>,
         reverse: IdVec<Atom, Elem>,
@@ -92,6 +138,7 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> Renumbering<Atom, Elem> {
         debug_assert!(Self::is_inverse(&forward, &reverse));
         Self::new_unchecked(forward, reverse, old_generation, new_generation)
     }
+    /// Returns the new variable corresponding to the given old variable, if it exists.
     pub fn old_to_new(&self, old: Elem) -> Option<Elem> {
         self.forward
             .get(old.atom())
@@ -99,9 +146,11 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> Renumbering<Atom, Elem> {
             .flatten()
             .map(|e| e.apply_pol_of(old))
     }
+    /// Returns the old variable corresponding to the given new variable, if it exists.
     pub fn new_to_old(&self, new: Elem) -> Option<Elem> {
         self.reverse.get(new.atom()).map(|&e| e.apply_pol_of(new))
     }
+    /// Returns `true` iff the given renumbering satisfies the constraint that equivalent variables are mapped identically.
     pub fn is_repr_reduction(&self, union_find: &UnionFind<Atom, Elem>) -> bool {
         union_find.lowest_unused_atom() <= self.forward.next_unused_key()
             && self.forward.iter().all(|(old, &new)| {
@@ -112,10 +161,15 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> Renumbering<Atom, Elem> {
     }
 }
 
+/// `Change` represents a single change of a `TrackedUnionFind`
+#[allow(missing_docs)] // dont want to document every subfield
 #[derive(Clone)]
 pub enum Change<Atom, Elem> {
+    /// A `union` operation. The set with representative `merged_repr` is merged into the set with representative `new_repr`.
     Union { new_repr: Atom, merged_repr: Elem },
+    /// A `make_repr` operation. `new_repr` is promoted to be the representative of its set, replacing `old_repr`.
     MakeRepr { new_repr: Atom, old_repr: Elem },
+    /// A renumbering operation.
     Renumber(Arc<Renumbering<Atom, Elem>>),
 }
 
@@ -140,6 +194,7 @@ impl<Atom: Id, Elem: Id> std::fmt::Debug for Change<Atom, Elem> {
     }
 }
 
+/// A `TrackedUnionFind` augments a [`UnionFind`] structure with change tracking.
 pub struct TrackedUnionFind<Atom, Elem> {
     tuf_id: TrackedUnionFindId,
     union_find: UnionFind<Atom, Elem>,
@@ -150,46 +205,13 @@ pub struct TrackedUnionFind<Atom, Elem> {
     generation: Generation,
 }
 
-pub struct IdAlloc<T> {
-    counter: Atomic<T>,
-}
-
-impl<T: Id + NoUninit> Default for IdAlloc<T> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<T: Id + NoUninit> IdAlloc<T> {
-    const fn new() -> Self {
-        Self {
-            counter: Atomic::new(T::MIN_ID),
-        }
-    }
-    pub fn alloc_block(&self, n: usize) -> T {
-        use atomic::Ordering::Relaxed;
-        debug_assert!(n > 0);
-        self.counter
-            .fetch_update(Relaxed, Relaxed, |current_id| {
-                current_id
-                    .id_index()
-                    .checked_add(n)
-                    .and_then(T::try_from_id_index)
-            })
-            .expect("not enough IDs remaining")
-    }
-    pub fn alloc(&self) -> T {
-        self.alloc_block(1)
-    }
-}
-
 static TUF_ID_ALLOC: IdAlloc<TrackedUnionFindId> = IdAlloc::new();
 
-impl<Atom, Elem> Default for TrackedUnionFind<Atom, Elem> {
-    fn default() -> Self {
+impl<Atom, Elem> From<UnionFind<Atom, Elem>> for TrackedUnionFind<Atom, Elem> {
+    fn from(union_find: UnionFind<Atom, Elem>) -> Self {
         Self {
-            tuf_id: TUF_ID_ALLOC.alloc(),
-            union_find: Default::default(),
+            union_find,
+            tuf_id: TUF_ID_ALLOC.alloc().unwrap(),
             log: Default::default(),
             observer_id_alloc: Default::default(),
             observers: Default::default(),
@@ -199,12 +221,31 @@ impl<Atom, Elem> Default for TrackedUnionFind<Atom, Elem> {
     }
 }
 
-impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> TrackedUnionFind<Atom, Elem> {
+impl<Atom, Elem> Default for TrackedUnionFind<Atom, Elem> {
+    fn default() -> Self {
+        UnionFind::default().into()
+    }
+}
+
+impl<Atom, Elem> TrackedUnionFind<Atom, Elem> {
+    /// Returns a shared reference to the contained `UnionFind`.
+    pub fn get_union_find(&self) -> &UnionFind<Atom, Elem> {
+        &self.union_find
+    }
+    /// Returns the contained `UnionFind`. All change tracking data is lost.
+    pub fn into_union_find(self) -> UnionFind<Atom, Elem> {
+        self.union_find
+    }
+}
+
+impl<Atom: Id, Elem: Id + Element<Atom>> TrackedUnionFind<Atom, Elem> {
+    /// Returns an element's representative. See [`UnionFind::find`].
     pub fn find(&self, elem: Elem) -> Elem {
         self.union_find.find(elem)
     }
-    pub fn union_full(&mut self, lits: [Elem; 2]) -> (bool, [Elem; 2]) {
-        let (ok, roots) = self.union_find.union_full(lits);
+    /// Declares two elements to be equivalent. See [`UnionFind::union_full`].
+    pub fn union_full(&mut self, elems: [Elem; 2]) -> (bool, [Elem; 2]) {
+        let (ok, roots) = self.union_find.union_full(elems);
         if ok && !self.observers.is_empty() {
             let new_repr = roots[0].atom();
             let merged_repr = roots[1].apply_pol_of(roots[0]);
@@ -215,9 +256,11 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> TrackedUnionFind<Atom, Elem>
         }
         (ok, roots)
     }
+    /// Declares two elements to be equivalent. See [`UnionFind::union`].
     pub fn union(&mut self, lits: [Elem; 2]) -> bool {
         self.union_full(lits).0
     }
+    /// Promotes an atom to be a representative. See [`UnionFind::make_repr`].
     pub fn make_repr(&mut self, new_repr: Atom) -> Elem {
         let old_repr = self.union_find.make_repr(new_repr);
         if old_repr.atom() != new_repr && !self.observers.is_empty() {
@@ -225,6 +268,12 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> TrackedUnionFind<Atom, Elem>
         }
         old_repr
     }
+    /// Renumbers all the variables in the `UnionFind`.
+    /// The provided mapping **must** meet all the preconditions listed for [`Renumbering`].
+    ///
+    /// This resets the `UnionFind` to the trivial state (`find(a) == a` for all `a`) and increments the generation ID.
+    ///
+    /// This method will panic in debug mode if said preconditions are not met.
     pub fn renumber(&mut self, forward: IdVec<Atom, Option<Elem>>, reverse: IdVec<Atom, Elem>) {
         let old_generation = self.generation;
         let new_generation = Generation(old_generation.0 + 1);
@@ -239,14 +288,23 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> TrackedUnionFind<Atom, Elem>
 }
 
 impl<Atom, Elem> TrackedUnionFind<Atom, Elem> {
+    /// Constructs a new, empty `TrackedUnionFind`.
     pub fn new() -> Self {
         Self::default()
     }
     fn log_end(&self) -> u64 {
         self.log_start + self.log.len() as u64
     }
+    /// Creates a new `ObserverToken` that can be used to track all changes since the call to this method.
+    ///
+    /// Conceptually, each observer has its own private log.
+    /// Any changes that happen to the `UnionFind` will be recorded into the logs of all currently active observers.
+    /// (In the actual implementation, only a single log is kept).
+    ///
+    /// After use, the `ObserverToken` must be disposed of with a call to `stop_observing`, otherwise
+    /// the memory corresponding to old log entries cannot be reclaimed until the `TrackedUnionFind` is dropped.
     pub fn start_observing(&mut self) -> ObserverToken {
-        let observer_id = self.observer_id_alloc.alloc();
+        let observer_id = self.observer_id_alloc.alloc().unwrap();
         self.observers.push(observer_id, Reverse(self.log_end()));
         ObserverToken {
             tuf_id: self.tuf_id,
@@ -254,9 +312,10 @@ impl<Atom, Elem> TrackedUnionFind<Atom, Elem> {
             observer_id,
         }
     }
+    /// Clones an `ObserverToken`, conceptually cloning the token's private log.
     pub fn clone_token(&mut self, token: &ObserverToken) -> ObserverToken {
         assert!(token.tuf_id == self.tuf_id);
-        let new_observer_id = self.observer_id_alloc.alloc();
+        let new_observer_id = self.observer_id_alloc.alloc().unwrap();
         let pos = *self.observers.get_priority(&token.observer_id).unwrap();
         self.observers.push(new_observer_id, pos);
         ObserverToken {
@@ -265,6 +324,9 @@ impl<Atom, Elem> TrackedUnionFind<Atom, Elem> {
             observer_id: new_observer_id,
         }
     }
+    /// Deletes an `ObserverToken` and its associated state.
+    ///
+    /// You must call this to allow the `TrackedUnionFind` to allow memory to be reclaimed.
     pub fn stop_observing(&mut self, token: ObserverToken) {
         assert!(token.tuf_id == self.tuf_id);
         self.observers.remove(&token.observer_id);
@@ -275,13 +337,11 @@ impl<Atom, Elem> TrackedUnionFind<Atom, Elem> {
             if new_start > self.log_start {
                 let delete = (new_start - self.log_start).try_into().unwrap();
                 drop(self.log.drain(0..delete));
-                println!("dropped {delete} entries");
                 self.log_start = new_start;
             }
         } else {
             self.log_start = self.log_end();
             self.log.clear();
-            println!("dropped all entries");
         }
     }
     fn observer_rel_pos(&self, token: &ObserverToken) -> usize {
@@ -307,6 +367,7 @@ impl<Atom, Elem> TrackedUnionFind<Atom, Elem> {
             .change_priority(&token.observer_id, Reverse(abs_pos));
         self.truncate_log();
     }
+    #[allow(clippy::type_complexity)]
     fn change_slices(
         &self,
         token: &ObserverToken,
@@ -331,6 +392,17 @@ impl<Atom, Elem> TrackedUnionFind<Atom, Elem> {
             }
         }
     }
+    /// Calls the provided function `f` with the content of the token's private log and clears the log.
+    ///
+    /// Because the log is not necessarily contiguous in memory, `f` may be called multiple times.
+    ///
+    /// The slice argument to `f` is guaranteed to be non-empty.
+    /// To allow looking up representatives `f` is also provided with a shared reference to the `UnionFind`.
+    ///
+    /// The method assumes that you will immediately process any `Renumbering` operations in the log
+    /// and will update the token's generation field.
+    /// 
+    /// Returns `true` iff `f` has been called at least once.
     pub fn drain_changes_with_fn(
         &mut self,
         token: &mut ObserverToken,
@@ -351,6 +423,11 @@ impl<Atom, Elem> TrackedUnionFind<Atom, Elem> {
             false
         }
     }
+    /// Returns a draining iterator that returns and deletes entries from the token's private log.
+    ///
+    /// Dropping this iterator will clear any unread entries, call `stop` if this is undesirable.
+    ///
+    /// You must not leak the returned iterator. Otherwise log entries may be observed multiple times and appear duplicated.
     pub fn drain_changes<'a>(
         &mut self,
         token: &'a mut ObserverToken,
@@ -365,21 +442,37 @@ impl<Atom, Elem> TrackedUnionFind<Atom, Elem> {
     }
 }
 
+/// A draining iterator.
+///
+/// Since this is a lending iterator, it does not implement the standard `Iterator` trait,
+/// but its `map` and `cloned` methods will create a standard iterator.
 pub struct DrainChanges<'a, 'b, Atom, Elem> {
     tuf: &'a mut TrackedUnionFind<Atom, Elem>,
     token: &'b mut ObserverToken,
     rel_pos: usize,
 }
 
+/// A draining iterator that has been mapped.
 pub struct DrainChangesMap<'a, 'b, Atom, Elem, F> {
     inner: DrainChanges<'a, 'b, Atom, Elem>,
     f: F,
 }
 
 impl<'a, 'b, Atom, Elem> DrainChanges<'a, 'b, Atom, Elem> {
+    /// Returns a reference to the first entry in the token's private log, without deleting it.
+    /// 
+    /// Returns `None` if the log is empty.
     pub fn peek(&mut self) -> Option<&Change<Atom, Elem>> {
         self.tuf.log.get(self.rel_pos)
     }
+    /// Returns a reference to the first entry in the token's private log. The entry will be deleted after its use.
+    /// 
+    /// If this returns a `Renumbering`, it is assumed that you will process it and the token's generation number will be updated.
+    /// 
+    /// Returns `None` if the log is empty. If `next` returned `None`, it will never return any more entries (the iterator is fused).
+    ///
+    /// As reflected by the lifetimes, the API only guarantees that the returned reference until the next call of any method of this iterator.
+    /// (In practice, deletion is more lazy and happens on drop).
     #[allow(clippy::should_implement_trait)]
     pub fn next(&mut self) -> Option<&Change<Atom, Elem>> {
         let ret = self.tuf.log.get(self.rel_pos);
@@ -390,14 +483,20 @@ impl<'a, 'b, Atom, Elem> DrainChanges<'a, 'b, Atom, Elem> {
         }
         ret
     }
+    /// Drops the iterator but without deleting unread entries.
     pub fn stop(self) {
         self.tuf.observer_set_rel_pos(self.token, self.rel_pos);
         let _ = ManuallyDrop::new(self);
     }
+    /// Returns `(n, Some(n))` where `n` is the number of unread entries.
+    ///
+    /// This method is designed to be compatible with the standard iterator method of the same name.
     pub fn size_hint(&self) -> (usize, Option<usize>) {
         let count = self.tuf.log.len() - self.rel_pos;
         (count, Some(count))
     }
+    /// Creates a new iterator by lazily calling `f` on every change.
+    #[must_use]
     pub fn map<B, F>(self, f: F) -> DrainChangesMap<'a, 'b, Atom, Elem, F>
     where
         F: FnMut(&Change<Atom, Elem>) -> B,
@@ -407,6 +506,9 @@ impl<'a, 'b, Atom, Elem> DrainChanges<'a, 'b, Atom, Elem> {
 }
 
 impl<'a, 'b, Atom: Clone, Elem: Clone> DrainChanges<'a, 'b, Atom, Elem> {
+    /// Create a standard iterator by cloning every entry.
+    #[must_use]
+    #[allow(clippy::type_complexity)]
     pub fn cloned(
         self,
     ) -> DrainChangesMap<'a, 'b, Atom, Elem, fn(&Change<Atom, Elem>) -> Change<Atom, Elem>> {
@@ -416,6 +518,9 @@ impl<'a, 'b, Atom: Clone, Elem: Clone> DrainChanges<'a, 'b, Atom, Elem> {
 
 impl<Atom, Elem> Drop for DrainChanges<'_, '_, Atom, Elem> {
     fn drop(&mut self) {
+        // mark any renumberings as seen
+        while self.next().is_some() {
+        }
         self.tuf
             .observer_set_rel_pos(self.token, self.tuf.log.len());
     }
@@ -435,33 +540,12 @@ where
     }
 }
 
-#[test]
-fn test() {
-    use imctk_lit::{Lit, Var};
-    let l = |n| Var::from_index(n).as_lit();
-    let mut tuf = TrackedUnionFind::<Var, Lit>::new();
-    let mut token = tuf.start_observing();
-    tuf.union([l(3), !l(4)]);
-    tuf.union([l(8), l(7)]);
-    let mut token2 = tuf.start_observing();
-    tuf.union([l(4), l(5)]);
-    for change in tuf.drain_changes(&mut token).cloned() {
-        println!("{change:?}");
-    }
-    println!("---");
-    tuf.union([!l(5), l(6)]);
-    tuf.make_repr(l(4).var());
-    let renumber: IdVec<Var, Option<Lit>> =
-        IdVec::from_vec(vec![Some(l(0)), None, None, Some(l(1)), Some(!l(1)), Some(!l(1)), Some(l(1)), Some(l(2)), Some(l(2))]);
-    let reverse = Renumbering::get_reverse(&renumber, &tuf.union_find);
-    dbg!(&renumber, &reverse);
-    tuf.renumber(renumber, reverse);
-    tuf.union([l(0), l(1)]);
-    let mut iter = tuf.drain_changes(&mut token);
-    println!("{:?}", iter.next());
-    iter.stop();
-    println!("---");
-    for change in tuf.drain_changes(&mut token2).cloned() {
-        println!("{change:?}");
-    }
+impl<Atom, Elem, B, F> ExactSizeIterator for DrainChangesMap<'_, '_, Atom, Elem, F> where
+    F: FnMut(&Change<Atom, Elem>) -> B
+{
+}
+
+impl<Atom, Elem, B, F> FusedIterator for DrainChangesMap<'_, '_, Atom, Elem, F> where
+    F: FnMut(&Change<Atom, Elem>) -> B
+{
 }
