+++ dsf/src/union_find.rs
@@ -1,41 +1,53 @@
-#![allow(missing_docs)]
+//! `UnionFind` efficiently tracks equivalences between variables.
+use crate::Element;
+use atomic::Atomic;
+use imctk_ids::{id_vec::IdVec, Id, IdRange};
 use std::sync::atomic::Ordering;
 
-use atomic::Atomic;
-use bytemuck::NoUninit;
-use imctk_ids::{id_vec::IdVec, Id};
-use imctk_lit::{Lit, Var};
+#[cfg(test)]
+#[path = "tests/test_union_find.rs"]
+mod test_union_find;
 
-pub trait Element<Atom> {
-    fn from_atom(atom: Atom) -> Self;
-    fn atom(self) -> Atom;
-    fn apply_pol_of(self, other: Self) -> Self;
-}
-
-impl<T> Element<T> for T {
-    fn from_atom(atom: T) -> Self {
-        atom
-    }
-    fn atom(self) -> T {
-        self
-    }
-    fn apply_pol_of(self, _other: T) -> Self {
-        self
-    }
-}
-
-impl Element<Var> for Lit {
-    fn from_atom(atom: Var) -> Self {
-        atom.as_lit()
-    }
-    fn atom(self) -> Var {
-        self.var()
-    }
-    fn apply_pol_of(self, other: Self) -> Self {
-        self ^ other.pol()
-    }
-}
-
+/// `UnionFind` efficiently tracks equivalences between variables.
+///
+/// Given "elements" of type `Elem`, this structure keeps track of any known equivalences between these elements.
+/// Equivalences are assumed to be *transitive*, i.e. if `x = y` and `y = z`, then `x = z` is assumed to be true, and in fact,
+/// automatically discovered by this structure.
+///
+/// Unlike a standard union find data structure, this version also keeps track of the *polarity* of elements.
+/// For example, you might always have pairs of elements `+x` and `-x` that are exact opposites of each other.
+/// If equivalences between `+x` and `-y` are then discovered, the structure also understands that `-x` and `+y` are similarly equivalent.
+///
+/// An element without its polarity is called an *atom*.
+/// The [`Element<Atom>`](Element) trait is required on `Elem` so that this structure can relate elements, atoms and polarities.
+///
+/// For each set of elements that are equivalent up to a polarity change, this structure keeps track of a *representative*.
+/// Each element starts out as its own representative, and if two elements are declared equivalent, the representative of one becomes the representative of both.
+/// The method `find` returns the representative for any element.
+///
+/// To declare two elements as equivalent, use the `union` or `union_full` methods, see their documentation for details on their use.
+///
+/// NB: Since this structure stores atoms in an `IdVec`, the atoms used should ideally be a contiguous set starting at `Atom::MIN_ID`.
+/// 
+/// ## Example ##
+/// ```
+/// use imctk_lit::{Var, Lit};
+/// use dsf::UnionFind;
+/// 
+/// let mut union_find: UnionFind<Var, Lit> = UnionFind::new();
+/// let lit = |n| Var::from_index(n).as_lit();
+/// 
+/// assert_eq!(union_find.find(lit(4)), lit(4));
+/// 
+/// union_find.union([lit(3), lit(4)]);
+/// assert_eq!(union_find.find(lit(4)), lit(3));
+/// 
+/// union_find.union([lit(1), !lit(2)]);
+/// union_find.union([lit(2), lit(3)]);
+/// assert_eq!(union_find.find(lit(1)), lit(1));
+/// assert_eq!(union_find.find(lit(4)), !lit(1));
+/// 
+/// ```
 pub struct UnionFind<Atom, Elem> {
     parent: IdVec<Atom, Atomic<Elem>>,
 }
@@ -48,7 +60,7 @@ impl<Atom, Elem> Default for UnionFind<Atom, Elem> {
     }
 }
 
-impl<Atom: Id, Elem: NoUninit> Clone for UnionFind<Atom, Elem> {
+impl<Atom: Id, Elem: Id> Clone for UnionFind<Atom, Elem> {
     fn clone(&self) -> Self {
         let new_parent = self
             .parent
@@ -62,36 +74,63 @@ impl<Atom: Id, Elem: NoUninit> Clone for UnionFind<Atom, Elem> {
     }
 }
 
-impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> UnionFind<Atom, Elem> {
+impl<Atom, Elem> UnionFind<Atom, Elem> {
+    /// Constructs an empty `UnionFind`.
+    ///
+    /// The returned struct obeys `find(a) == a` for all `a`.
     pub fn new() -> Self {
         UnionFind::default()
     }
+}
+
+impl<Atom: Id, Elem: Id + Element<Atom>> UnionFind<Atom, Elem> {
+    /// Constructs an empty `UnionFind`, with room for `capacity` elements.
+    pub fn with_capacity(capacity: usize) -> Self {
+        UnionFind {
+            parent: IdVec::from_vec(Vec::with_capacity(capacity)),
+        }
+    }
+    /// Clears all equivalences, but retains any allocated memory.
+    pub fn clear(&mut self) {
+        self.parent.clear();
+    }
     fn read_parent(&self, atom: Atom) -> Elem {
-        self.parent
-            .get(atom)
-            .map(|p| p.load(Ordering::Relaxed))
-            .unwrap_or(Elem::from_atom(atom))
+        if let Some(parent_cell) = self.parent.get(atom) {
+            // This load is allowed to reorder with stores from `update_parent`, see there for details.
+            parent_cell.load(Ordering::Relaxed)
+        } else {
+            Elem::from_atom(atom)
+        }
     }
+    // Important: Only semantically trivial changes are allowed using this method!!
+    // Specifically, update_parent(atom, parent) should only be called if `parent` is already an ancestor of `atom`
+    // Otherwise, concurrent calls to `read_parent` (which are explicitly allowed!) could return incorrect results.
     fn update_parent(&self, atom: Atom, parent: Elem) {
-        let Some(parent_ref) = self.parent.get(atom) else {
+        if let Some(parent_cell) = self.parent.get(atom) {
+            parent_cell.store(parent, Ordering::Relaxed);
+        } else {
+            // can only get here if the precondition or a data structure invariant is violated
             panic!("shouldn't happen: update_parent called with out of bounds argument");
-        };
-        parent_ref.store(parent, Ordering::Relaxed);
+        }
     }
+    // Unlike `update_parent`, this is safe for arbitrary updates, since it requires &mut self.
     fn write_parent(&mut self, atom: Atom, parent: Elem) {
-        if let Some(parent_ref) = self.parent.get(atom) {
-            parent_ref.store(parent, Ordering::Relaxed);
+        if let Some(parent_cell) = self.parent.get(atom) {
+            parent_cell.store(parent, Ordering::Relaxed);
         } else {
             debug_assert!(self.parent.next_unused_key() <= atom);
             while self.parent.next_unused_key() < atom {
-                self.parent
-                    .push(Atomic::new(Elem::from_atom(self.parent.next_unused_key())));
+                let next_elem = Elem::from_atom(self.parent.next_unused_key());
+                self.parent.push(Atomic::new(next_elem));
             }
             self.parent.push(Atomic::new(parent));
         }
     }
     fn find_root(&self, mut elem: Elem) -> Elem {
         loop {
+            // If we interleave with a call to `update_parent`, the parent may change
+            // under our feet, but it's okay because in that case we get an ancestor instead,
+            // which just skips some iterations of the loop!
             let parent = self.read_parent(elem.atom()).apply_pol_of(elem);
             if elem == parent {
                 return elem;
@@ -100,220 +139,103 @@ impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> UnionFind<Atom, Elem> {
             elem = parent;
         }
     }
+    // Worst-case `find` performance is linear. To keep amortised time complexity logarithmic,
+    // we memoise the result of `find_root` by calling `update_parent` on every element
+    // we traversed.
     fn update_root(&self, mut elem: Elem, root: Elem) {
+        // Loop invariant: `root` is the representative of `elem`.
         loop {
+            // Like in `find_root`, this may interleave with `update_root` calls, and we may skip some steps,
+            // which is okay because the other thread will do the updates instead.
             let parent = self.read_parent(elem.atom()).apply_pol_of(elem);
             if parent == root {
                 break;
             }
+            // By the loop invariant, this just sets `elem`'s parent to its representative,
+            // which satisfies the precondition for `update_parent`. Further if two threads end up
+            // here simultaneously, they will both set to the same representative,
+            // therefore the change is idempotent.
             self.update_parent(elem.atom(), root.apply_pol_of(elem));
             elem = parent;
         }
     }
-    pub fn find(&self, lit: Elem) -> Elem {
-        let root = self.find_root(lit);
-        self.update_root(lit, root);
+    /// Returns the representative for an element. Elements are equivalent iff they have the same representative.
+    ///
+    /// Elements `a` and `b` are equivalent up to a polarity change iff they obey `find(a) = find(b) ^ p` for some polarity `p`.
+    ///
+    /// This operation is guaranteed to return `elem` itself for arguments `elem >= lowest_unused_atom()`.
+    ///
+    /// The amortised time complexity of this operation is **O**(log N).
+    pub fn find(&self, elem: Elem) -> Elem {
+        let root = self.find_root(elem);
+        self.update_root(elem, root);
         root
     }
-    pub fn union_full(&mut self, lits: [Elem; 2]) -> (bool, [Elem; 2]) {
-        let [a, b] = lits;
+    /// Declares two elements to be equivalent. The new representative of both is the representative of the first element.
+    ///
+    /// If the elements are already equivalent or cannot be made equivalent (are equivalent up to a sign change),
+    /// the operation returns `false` without making any changes. Otherwise it returns `true`.
+    ///
+    /// In both cases it also returns the original representatives of both arguments.
+    ///
+    /// The amortised time complexity of this operation is **O**(log N).
+    pub fn union_full(&mut self, elems: [Elem; 2]) -> (bool, [Elem; 2]) {
+        let [a, b] = elems;
         let ra = self.find(a);
         let rb = self.find(b);
         if ra.atom() == rb.atom() {
             (false, [ra, rb])
         } else {
+            // The first write is only needed to ensure that the parent table actually contains `a`
+            // and is a no-op otherwise.
+            self.write_parent(ra.atom(), Elem::from_atom(ra.atom()));
             self.write_parent(rb.atom(), ra.apply_pol_of(rb));
             (true, [ra, rb])
         }
     }
-    pub fn union(&mut self, lits: [Elem; 2]) -> bool {
-        self.union_full(lits).0
+    /// Declares two elements to be equivalent. The new representative of both is the representative of the first element.
+    ///
+    /// If the elements are already equivalent or cannot be made equivalent (are equivalent up to polarity),
+    /// the operation returns `false` without making any changes. Otherwise it returns `true`.
+    ///
+    /// The amortised time complexity of this operation is **O**(log N).
+    pub fn union(&mut self, elems: [Elem; 2]) -> bool {
+        self.union_full(elems).0
     }
+    /// Sets `atom` to be its own representative, and updates other representatives to preserve all existing equivalences.
+    ///
+    /// The amortised time complexity of this operation is **O**(log N).
     pub fn make_repr(&mut self, atom: Atom) -> Elem {
         let root = self.find(Elem::from_atom(atom));
         self.write_parent(atom, Elem::from_atom(atom));
         self.write_parent(root.atom(), Elem::from_atom(atom).apply_pol_of(root));
         root
     }
+    /// Returns the lowest `Atom` value for which no equivalences are known.
+    ///
+    /// It is guaranteed that `find(a) == a` if `a >= lowest_unused_atom`, but the converse may not hold.
     pub fn lowest_unused_atom(&self) -> Atom {
         self.parent.next_unused_key()
     }
+    /// Returns an iterator that yields all tracked atoms and their representatives.
+    pub fn iter(&self) -> impl '_ + Iterator<Item = (Atom, Elem)> {
+        IdRange::from(Atom::MIN_ID..self.lowest_unused_atom())
+            .iter()
+            .map(|atom| (atom, self.find(Elem::from_atom(atom))))
+    }
 }
 
-#[cfg(test)]
-#[allow(dead_code)]
-mod tests {
-    use super::*;
-    use imctk_ids::id_set_seq::IdSetSeq;
-    use rand::prelude::*;
-    use std::collections::{HashSet, VecDeque};
-
-    #[derive(Default)]
-    struct CheckedUnionFind<Atom: Id, Elem> {
-        dut: UnionFind<Atom, Elem>,
-        equivs: IdSetSeq<Atom, Elem>,
-    }
-
-    impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> UnionFind<Atom, Elem> {
-        fn debug_print_tree(
-            children: &IdVec<Atom, Vec<Elem>>,
-            atom: Atom,
-            prefix: &str,
-            self_char: &str,
-            further_char: &str,
-            pol: bool,
-        ) {
-            println!(
-                "{prefix}{self_char}{}{:?}",
-                if pol { "!" } else { "" },
-                atom
-            );
-            let my_children = children.get(atom).unwrap();
-            for (index, &child) in my_children.iter().enumerate() {
-                let last = index == my_children.len() - 1;
-                let self_char = if last { "└" } else { "├" };
-                let next_further_char = if last { " " } else { "│" };
-                Self::debug_print_tree(
-                    children,
-                    child.atom(),
-                    &(prefix.to_string() + further_char),
-                    self_char,
-                    next_further_char,
-                    pol ^ (child != Elem::from_atom(child.atom())),
-                );
+impl<Atom: Id, Elem: Id + Element<Atom>> std::fmt::Debug for UnionFind<Atom, Elem> {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        // prints non-trivial sets of equivalent elements, always printing the representative first
+        let mut sets = std::collections::HashMap::<Atom, Vec<Elem>>::new();
+        for (atom, repr) in self.iter() {
+            if Elem::from_atom(atom) != repr {
+                sets.entry(repr.atom())
+                    .or_insert_with(|| vec![Elem::from_atom(repr.atom())])
+                    .push(Elem::from_atom(atom).apply_pol_of(repr));
             }
         }
-        fn debug_print(&self) {
-            let mut children: IdVec<Atom, Vec<Elem>> = Default::default();
-            for atom in self.parent.keys() {
-                let parent = self.read_parent(atom);
-                children.grow_for_key(atom);
-                if atom != parent.atom() {
-                    children
-                        .grow_for_key(parent.atom())
-                        .push(Elem::from_atom(atom).apply_pol_of(parent));
-                } else {
-                    assert!(Elem::from_atom(atom) == parent);
-                }
-            }
-            for atom in self.parent.keys() {
-                if atom == self.read_parent(atom).atom() {
-                    Self::debug_print_tree(&children, atom, "", "", " ", false);
-                }
-            }
-        }
-    }
-    #[derive(Debug, Copy, Clone, PartialOrd, Ord, PartialEq, Eq)]
-    enum VarRel {
-        Equiv,
-        AntiEquiv,
-        NotEquiv,
-    }
-
-    impl<Atom: Id, Elem: Id + Element<Atom> + NoUninit> CheckedUnionFind<Atom, Elem> {
-        fn new() -> Self {
-            CheckedUnionFind {
-                dut: Default::default(),
-                equivs: Default::default(),
-            }
-        }
-        fn ref_equal(&mut self, start: Elem, goal: Elem) -> VarRel {
-            let mut seen: HashSet<Atom> = Default::default();
-            let mut queue: VecDeque<Elem> = [start].into();
-            while let Some(place) = queue.pop_front() {
-                if place.atom() == goal.atom() {
-                    if place == goal {
-                        return VarRel::Equiv;
-                    } else {
-                        return VarRel::AntiEquiv;
-                    }
-                }
-                seen.insert(place.atom());
-                for &next in self.equivs.grow_for(place.atom()).iter() {
-                    if !seen.contains(&next.atom()) {
-                        queue.push_back(next.apply_pol_of(place));
-                    }
-                }
-            }
-            VarRel::NotEquiv
-        }
-        fn find(&mut self, lit: Elem) -> Elem {
-            let out = self.dut.find(lit);
-            assert!(self.ref_equal(lit, out) == VarRel::Equiv);
-            out
-        }
-        fn union_full(&mut self, lits: [Elem; 2]) -> (bool, [Elem; 2]) {
-            let (ok, [ra, rb]) = self.dut.union_full(lits);
-            assert_eq!(self.ref_equal(lits[0], ra), VarRel::Equiv);
-            assert_eq!(self.ref_equal(lits[1], rb), VarRel::Equiv);
-            assert_eq!(ok, self.ref_equal(lits[0], lits[1]) == VarRel::NotEquiv);
-            assert_eq!(self.dut.find_root(lits[0]), ra);
-            if ok {
-                assert_eq!(self.dut.find_root(lits[1]), ra);
-                self.equivs
-                    .grow_for(lits[0].atom())
-                    .insert(lits[1].apply_pol_of(lits[0]));
-                self.equivs
-                    .grow_for(lits[1].atom())
-                    .insert(lits[0].apply_pol_of(lits[1]));
-            } else {
-                assert_eq!(self.dut.find_root(lits[1]).atom(), ra.atom());
-            }
-            (ok, [ra, rb])
-        }
-        fn union(&mut self, lits: [Elem; 2]) -> bool {
-            self.union_full(lits).0
-        }
-        fn make_repr(&mut self, lit: Atom) {
-            self.dut.make_repr(lit);
-            assert_eq!(
-                self.dut.find_root(Elem::from_atom(lit)),
-                Elem::from_atom(lit)
-            );
-            self.check();
-        }
-        fn check(&mut self) {
-            for atom in self.dut.parent.keys() {
-                let parent = self.dut.read_parent(atom);
-                assert_eq!(self.ref_equal(Elem::from_atom(atom), parent), VarRel::Equiv);
-                let root = self.dut.find_root(Elem::from_atom(atom));
-                for &child in self.equivs.grow_for(atom).iter() {
-                    assert_eq!(root, self.dut.find_root(child));
-                }
-            }
-        }
-    }
-
-    #[test]
-    fn test() {
-        let mut u: CheckedUnionFind<Var, Lit> = CheckedUnionFind::new();
-        let mut rng = rand_pcg::Pcg64::seed_from_u64(25);
-        let max_var = 2000;
-        for i in 0..2000 {
-            match rng.gen_range(0..10) {
-                0..=4 => {
-                    let a = Lit::from_code(rng.gen_range(0..=2 * max_var + 1));
-                    let b = Lit::from_code(rng.gen_range(0..=2 * max_var + 1));
-                    let result = u.union_full([a, b]);
-                    println!("union({a}, {b}) = {result:?}");
-                }
-                5..=7 => {
-                    let a = Lit::from_code(rng.gen_range(0..=2 * max_var + 1));
-                    let result = u.find(a);
-                    println!("find({a}) = {result}");
-                }
-                8 => {
-                    u.check();
-                }
-                9 => {
-                    let a = Var::from_index(rng.gen_range(0..=max_var));
-                    u.make_repr(a);
-                    println!("make_repr({a})");
-                }
-                _ => {}
-            }
-        }
-        u.check();
-        //u.dut.debug_print();
+        f.debug_set().entries(sets.values()).finish()
     }
 }
